+++ Math/easymath.h
@@ -42,7 +42,7 @@ double rand(double low, double high);
 //! Error function (this exists in linux but not windows)
 double erfc(double x);
 
-int get_nearest_square(int n);
+size_t get_nearest_square(size_t n);
 
 std::set<XY> get_n_unique_points(double xmin, double xmax,
     double ymin, double ymax, size_t n);
@@ -52,13 +52,13 @@ std::vector<XY> get_n_unique_square_points(double xmin, double xmax,
     double ymin, double ymax, size_t n);
 
 //! Gets unique indices for a square that must contain n points
-std::vector<std::pair<int, int> > get_n_unique_square_subscripts(size_t n);
+std::vector<std::pair<size_t, size_t> > get_n_unique_square_subscripts(size_t n);
 
 std::pair<int, int> ind2sub(const int sub, const int cols, const int rows);
 
 size_t nChoosek(size_t n, size_t k);
 
-std::vector<std::pair<int, int> > all_combos_of_2(size_t n);
+std::vector<std::pair<size_t, size_t> > all_combos_of_2(size_t n);
 
 bool is_endpt(const XY& a, const line_segment& b);
 bool pt_on_line(const XY& a, const line_segment& b, const double thresh = 0.01);
